Refactor: styles.ts 파일을 페이지별로 분리
