+++ src/components/loginPage/index.ts
@@ -2,3 +2,4 @@ export { default as BackgroundElements } from './BackgroundElements'
 export { default as BrandingSection } from './BrandingSection'
 export { default as LoginForm } from './LoginForm'
 export { default as Header } from './Header'
+export { styles, cn } from './styles'
